SDK-176 #comment Changed calculation of job status as discussed on 2019-10-29

+++ njams-sdk/src/test/java/com/im/njams/sdk/logmessage/ActivityImplTest.java
@@ -79,7 +79,7 @@ public class ActivityImplTest extends AbstractTest {
 
         //Here the EventStatus is set to null, but the JobStatus is not affected.
         //The JobStatus is still SUCCESS.
-        act.setEventStatus(null);
+        act.setEventStatus((Integer) null);
         assertEquals(null, act.getEventStatus());
         assertEquals(JobStatus.SUCCESS, job.getStatus());
